moved vault relevant classes into separate namespace
